refacto: use auth reducer
