+++ App.tsx
@@ -23,6 +23,7 @@ import { ThemeProvider } from './src/shared/theme.context';
 import { AuthSubject } from './src/models/session.model';
 import SettingsScreen from './src/screens/settings/settings.screen';
 import ColorSelectScreen from './src/screens/settings/color-select.screen';
+import { authReducer } from './src/redux/reducers/auth.reducer';
 
 const RootStack = createStackNavigator();
 const SettingsStack = createStackNavigator();
@@ -51,68 +52,8 @@ export default function App() {
 
   const { inviteLink } = useInitialUrl();
 
-  const createAnonymousToken = async () => {
-    const values = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
-    let anonymousToken = '';
-
-    // eslint-disable-next-line no-plusplus
-    for (let i = 20; i > 0; --i) {
-      anonymousToken += values[Math.round(Math.random() * (values.length - 1))];
-    }
-
-    const session: AuthSubject = {
-      _id: `user-${anonymousToken}`,
-      name: null,
-      mail: null,
-      settings: {
-        theme: 'dark',
-        picture: null,
-        color: '#DF62A9',
-        isColorblind: false,
-      },
-    };
-
-    await AsyncStorage.setItem('BBOX-user', JSON.stringify(session));
-
-    return session;
-  };
-
   const [state, dispatch] = useReducer(
-    (prevState, action) => {
-      switch (action.type) {
-        case 'RESTORE_TOKEN':
-
-          if (action.token) {
-            axios.defaults.headers.common.Authorization = `Bearer ${action.token}`;
-          } else {
-            delete axios.defaults.headers.common.Authorization;
-            createAnonymousToken();
-          }
-
-          return {
-            ...prevState,
-            userToken: action.token,
-            isLoading: false,
-          };
-        case 'SIGN_IN':
-          axios.defaults.headers.common.Authorization = `Bearer ${action.token}`;
-          return {
-            ...prevState,
-            isSignout: false,
-            userToken: action.token,
-          };
-        case 'SIGN_OUT':
-          delete axios.defaults.headers.common.Authorization;
-          createAnonymousToken();
-          return {
-            ...prevState,
-            isSignout: true,
-            userToken: null,
-          };
-        default:
-          return {};
-      }
-    },
+    authReducer,
     {
       isLoading: true,
       isSignout: false,
